+++ NEWS.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-* Add `derefSymlinks` option (#410)
+* `derefSymlinks` option (#410)
 
 ## [7.1.0] - 2016-06-22