+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [7.2.0] - 2016-07-03
+
 ### Added
 
 * `derefSymlinks` option (#410)
@@ -167,6 +169,7 @@
 
 For versions prior to 5.2.0, please see `git log`.
 
+[7.2.0]: https://github.com/electron-userland/electron-packager/compare/v7.1.0...v7.2.0
 [7.1.0]: https://github.com/electron-userland/electron-packager/compare/v7.0.4...v7.1.0
 [7.0.4]: https://github.com/electron-userland/electron-packager/compare/v7.0.3...v7.0.4
 [7.0.3]: https://github.com/electron-userland/electron-packager/compare/v7.0.2...v7.0.3